+++ src/mem/n64bus.h
@@ -172,7 +172,7 @@ u32 n64_read_physical_word(u32 address);
 void n64_write_physical_half(u32 address, u16 value);
 u16 n64_read_physical_half(u32 address);
 
-void n64_write_physical_byte(u32 address, u8 value);
+void n64_write_physical_byte(u32 address, u32 value);
 u8 n64_read_physical_byte(u32 address);
 
 INLINE void n64_write_word(u64 address, u32 value) {
